Add IBowlingCentersApi and relocate summary model
Introduce IBowlingCentersApi Refit contract with ListBowlingCentersAsync (GET /bowling-centers). Move BowlingCenterSummaryResponse into the ListBowlingCenters namespace/path and update related using directives in endpoints and summary files. Register IBowlingCentersApi in ApiServicesConfiguration DI registration and update the test factory to use the new namespace.
