+++ tests/Neba.TestFactory/BowlingCenters/BowlingCenterSummaryResponseFactory.cs
@@ -1,6 +1,6 @@
 using Bogus;
 
-using Neba.Api.Contracts.BowlingCenters;
+using Neba.Api.Contracts.BowlingCenters.ListBowlingCenters;
 using Neba.Api.Contracts.Contact;
 using Neba.Application.Contact;
 using Neba.Domain.BowlingCenters;
